+++ includes/Header/Body_Class.php
@@ -5,7 +5,8 @@
  * @package    Michelle
  * @copyright  WebMan Design, Oliver Juhas
  *
- * @since  1.0.0
+ * @since    1.0.0
+ * @version  1.0.6
  */
 
 namespace WebManDesign\Michelle\Header;
@@ -120,7 +121,8 @@ class Body_Class implements Component_Interface {
 	/**
 	 * HTML body classes in admin area.
 	 *
-	 * @since  1.0.0
+	 * @since    1.0.0
+	 * @version  1.0.6
 	 *
 	 * @param  string $classes
 	 *
@@ -164,11 +166,12 @@ class Body_Class implements Component_Interface {
 
 			// "Editing footer" reusable block class.
 
-				if (
-					'wp_block' === get_post_type( $post )
-					&& get_the_ID( $post ) == get_theme_mod( 'block_area_site_footer' )
-				) {
-					$classes .= ' editing-block-area-site-footer';
+				if ( 'wp_block' === get_post_type( $post ) ) {
+					if ( get_the_ID( $post ) == get_theme_mod( 'block_area_site_footer' ) ) {
+						$classes .= ' editing-block-area-footer';
+					} else if ( get_the_ID( $post ) == get_theme_mod( 'block_area_error_404' ) ) {
+						$classes .= ' editing-block-area-404';
+					}
 				}