+++ changelog.md
@@ -1,5 +1,45 @@
 # Michelle Changelog
 
+## 1.0.6, 20210330
+
+### Added
+- Customize error 404 page content
+
+### Updated
+- Optimized fonts loading
+- Posts list styles
+- Site title styles
+- Starter content
+- Localization
+
+### File updates
+	404.php
+	changelog.md
+	readme.txt
+	style.css
+	assets/js/customize-controls.js
+	assets/scss/blocks.scss
+	assets/scss/content.scss
+	assets/scss/editor-style-blocks.scss
+	assets/scss/global.scss
+	includes/Autoload.php
+	includes/Content/Block_Area.php
+	includes/Content/Component.php
+	includes/Customize/Options.php
+	includes/Footer/Component.php
+	includes/Header/Body_Class.php
+	includes/Tool/Google_Fonts.php
+	languages/*.*
+	templates/parts/admin/content-starter-about.php
+	templates/parts/admin/content-starter-contact.php
+	templates/parts/admin/content-starter-faq.php
+	templates/parts/admin/content-starter-home.php
+	templates/parts/admin/content-starter-services.php
+	templates/parts/block/area-404.php
+	templates/parts/block/area-footer.php
+	templates/parts/content/content-404.php
+
+
 ## 1.0.5, 20210328
 
 ### Updated
@@ -16,20 +56,20 @@
 	changelog.md
 	readme.txt
 	style.css
-	assets\scss\blocks.scss
-	assets\scss\comments.scss
-	assets\scss\content.scss
-	assets\scss\editor-style-blocks.scss
-	assets\scss\editor-style-classic.scss
-	assets\scss\global.scss
-	includes\Autoload.php
-	includes\Content\Starter.php
-	includes\Customize\Options.php
-	includes\Customize\RGBA.php
-	includes\Plugin\Component.php
+	assets/scss/blocks.scss
+	assets/scss/comments.scss
+	assets/scss/content.scss
+	assets/scss/editor-style-blocks.scss
+	assets/scss/editor-style-classic.scss
+	assets/scss/global.scss
+	includes/Autoload.php
+	includes/Content/Starter.php
+	includes/Customize/Options.php
+	includes/Customize/RGBA.php
+	includes/Plugin/Component.php
 	languages/*.*
-	templates\parts\block\pattern-gallery-captions.php
-	templates\parts\block\pattern-team-2.php
+	templates/parts/block/pattern-gallery-captions.php
+	templates/parts/block/pattern-team-2.php
 
 
 ## 1.0.4, 20210327